refactor : API 관련 유틸 클래스 패키지 이동
