+++ src/main/java/com/pullanner/api/ApiResponseCode.java
@@ -1,4 +1,4 @@
-package com.pullanner.global.api;
+package com.pullanner.api;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
