+++ src/main/java/com/pullanner/global/exception/GlobalExceptionHandler.java
@@ -1,8 +1,8 @@
 package com.pullanner.global.exception;
 
-import com.pullanner.global.api.ApiResponseCode;
-import com.pullanner.global.api.ApiResponseMessage;
-import com.pullanner.global.api.ApiUtil;
+import com.pullanner.api.ApiResponseCode;
+import com.pullanner.api.ApiResponseMessage;
+import com.pullanner.api.ApiUtil;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
